requirements.txt
aiohttp==3.4.4
apipkg==1.5
argh==0.26.2
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.1
Babel==2.6.0
backcall==0.1.0
bleach==3.0.2
blessings==1.7
blinker==1.4
bokeh==1.0.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.4.1
cycler==0.10.0
daal==2019.0
decorator==4.3.0
defusedxml==0.5.0
dlib==19.16.99
docopt==0.6.2
docutils==0.14
doit==0.31.1
entrypoints==0.2.3
execnet==1.5.0
expects==0.8.0
face-recognition==1.2.3
face-recognition-models==0.3.0
flake8==3.5.0
future==0.17.1
ghp-import2==1.0.1
glob2==0.6
gpustat==0.5.0
graphviz==0.9
icc-rt==2019.0
idna==2.7
idna-ssl==1.1.0
intel-openmp==2019.0
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Logbook==1.4.1
lxml==4.2.5
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.4
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1.1
more-itertools==4.3.0
multidict==4.4.2
nano==2018.11.19
natsort==5.4.1
nbconvert==5.4.0
nbformat==4.4.0
-e git+git@github.com:necromuralist/In-Too-Deep.git@47d862413b2dba40e28389b28c92d4cd410b8388#egg=neurotic
Nikola==8.0.1
notebook==5.7.0
numpy==1.15.3
nvidia-ml-py3==7.352.0
opencv-python==3.4.4.19
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parse==1.9.0
parse-type==0.4.2
parso==0.3.1
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.5
piexif==1.0.13
Pillow==5.3.0
pip-tools==3.1.0
pipdeptree==0.13.1
pkg-resources==0.0.0
pluggy==0.8.0
prometheus-client==0.4.2
prompt-toolkit==2.0.6
psutil==5.4.8
ptpython==2.0.3
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pydaal==2019.0.0.20180713
pyflakes==1.6.0
Pygments==2.2.0
pyinotify==0.9.6
pyparsing==2.2.2
PyRSS2Gen==1.1
pyte==0.8.0
pytest==3.9.1
pytest-bdd==3.0.0
pytest-forked==0.2
pytest-mock==1.10.0
pytest-xdist==1.24.0
python-dateutil==2.7.3
python-dotenv==0.9.1
pyttsx3==2.7
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.20.0
requests-toolbelt==0.8.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tabulate==0.8.2
tbb==2019.0
tbb4py==2019.0
terminado==0.8.1
testpath==0.4.2
thefuck==3.28
torch==0.4.1
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
udacity-pa==0.2.8
Unidecode==1.0.22
urllib3==1.24
virtualenv==16.0.0
virtualfish==1.0.6
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
yapf==0.24.0
Yapsy==1.12.0
yarl==1.2.6
